+++ crates/nostr/src/nips/nip57.rs
@@ -31,8 +31,8 @@ use crate::types::time::TimeSupplier;
 #[cfg(feature = "std")]
 use crate::SECP256K1;
 use crate::{
-    event, util, Event, EventBuilder, EventId, JsonUtil, Keys, Kind, PublicKey, SecretKey, Tag,
-    TagStandard, Timestamp, Url,
+    event, util, Event, EventBuilder, EventId, JsonUtil, Keys, Kind, PublicKey, RelayUrl,
+    SecretKey, Tag, TagStandard, Timestamp,
 };
 
 type Aes256CbcEnc = Encryptor<Aes256>;
@@ -126,7 +126,7 @@ pub struct ZapRequestData {
     /// Public key of the recipient
     pub public_key: PublicKey,
     /// List of relays the recipient's wallet should publish its zap receipt to
-    pub relays: Vec<Url>,
+    pub relays: Vec<RelayUrl>,
     /// Message
     pub message: String,
     /// Amount in `millisats` the sender intends to pay
@@ -143,7 +143,7 @@ impl ZapRequestData {
     /// New Zap Request Data
     pub fn new<I>(public_key: PublicKey, relays: I) -> Self
     where
-        I: IntoIterator<Item = Url>,
+        I: IntoIterator<Item = RelayUrl>,
     {
         Self {
             public_key,
@@ -425,7 +425,7 @@ mod tests {
         let alice_keys = Keys::generate();
         let bob_keys = Keys::generate();
 
-        let relays = [Url::parse("wss://relay.damus.io").unwrap()];
+        let relays = [RelayUrl::parse("wss://relay.damus.io").unwrap()];
         let msg = "Private Zap message!";
         let data = ZapRequestData::new(bob_keys.public_key(), relays).message(msg);
         let private_zap = private_zap_request(data, &alice_keys).unwrap();